fix: 创建smart相关mapper，service并将相关user操作移动到samrt相关mapper、service中。分离smart、user
