+++ backend/src/main/java/com/fall/smarthouse/mapper/SmartMapper.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.fall.smarthouse.mapper.SmartMapper">
+
+</mapper>
