+++ backend/src/main/java/com/fall/smarthouse/service/IUserService.java
@@ -29,33 +29,6 @@ public interface IUserService {
     void sendEmail(String subject,String text,String account);
 
     /**
-     * @author FAll
-     * @description 网页获取侧边栏路由信息
-     * @return: java.util.ArrayList<com.fall.smarthouse.bean.MenuItem>
-     * @date 2022/12/3 10:17
-     */
-    ArrayList<MenuItem> getMenu();
-
-    /**
-     * @author FAll
-     * @description 用户登录
-     * @param account
-     * @param password
-     * @return: java.lang.Boolean
-     * @date 2022/12/6 13:36
-     */
-    String userLogin(String account,String password);
-
-    /**
-     * @author FAll
-     * @description 校验Token是否过期
-     * @param token
-     * @return: java.lang.Boolean
-     * @date 2022/12/6 16:02
-     */
-    Boolean checkLogin(String token) ;
-
-    /**
      * @description: 获取所有邮箱
      * @author xiaoQe
      * @date 2022/12/28 19:38
